[ClientApp] renaming Classes and anywhere they are called
